backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
groq==0.9.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.0
python-pptx==0.6.21
redis==5.0.1